Refactor the decrypt function
This allows for the decryption of the Dutch eHerkenning SAML implementation.
The logic is as follows:

Find all the EncryptedKey nodes in the XML and loop over these nodes. Because
the encrypted data belongs to the same grandparent
(keyNode->parentNode->parentNode) we can get the encrypted node and decrypt it
based on the key. That is if it is signed against our key.

A similar approach is taken for nodes which have a RetrievalMethod element.

Signed-off-by: Wesley Schwengle <waterkip@cpan.org>